refactor: enhance image handling and component styles across multiple sections
- Integrated the `imageSize` utility to optimize image loading in components such as CarShowcase, CarFleetImage, CarsHero, and TeamMember.
- Updated image classes to include `object-center` for better alignment.
- Removed deprecated fleet timeline data and streamlined car data imports for improved maintainability.
- Enhanced layout and responsiveness in various components, ensuring a consistent user experience.

These changes improve image performance and overall component styling, contributing to a more polished interface.
